Updated build path and added CLI_Driver.
CLI_Driver is just a text interface to what the program is doing in
terms of message passing and parsing so that we can interact without a
GUI for now.

I also updated my build path on my local copy of Eclipse that way it
could actually build successfully. Eclipse was still looking for a bunch
of libraries that existed, but had Patrick's build path as the default.
I replaced that on mine and I think before I merge this branch with the
master branch again, I will update .gitignore to include the build path
for Eclipse so that merges and forks won't affect our build paths. I'll
make a dedicated commit to .gitignore if I do so.