Added dialog class and basic Client GUI
I added a basic GUI for the client so that we can test and implement
some graphical functionality. I also added a simple custom dialog class
so that we can have easy server/port input from the client. I'm still
mulling over just using SWT/Swing for that because the way I did it
feels hackish.